+++ src/main/java/com/troblecodings/tcredstone/init/GIRCInit.java
@@ -31,40 +31,40 @@ import net.minecraftforge.registries.RegistryObject;
 
 public class GIRCInit {
 
-    public static final DeferredRegister<Item> ITEM_REGISTRY = DeferredRegister
-            .create(ForgeRegistries.ITEMS, GIRCRedstoneMain.MODID);
-    public static final DeferredRegister<Block> BLOCK_REGISTRY = DeferredRegister
-            .create(ForgeRegistries.BLOCKS, GIRCRedstoneMain.MODID);
-    public static final DeferredRegister<BlockEntityType<?>> TILEENTITY_REGISTRY = DeferredRegister
-            .create(ForgeRegistries.BLOCK_ENTITY_TYPES, GIRCRedstoneMain.MODID);
+    public static final DeferredRegister<Item> ITEM_REGISTRY =
+            DeferredRegister.create(ForgeRegistries.ITEMS, GIRCRedstoneMain.MODID);
+    public static final DeferredRegister<Block> BLOCK_REGISTRY =
+            DeferredRegister.create(ForgeRegistries.BLOCKS, GIRCRedstoneMain.MODID);
+    public static final DeferredRegister<BlockEntityType<?>> TILEENTITY_REGISTRY =
+            DeferredRegister.create(ForgeRegistries.BLOCK_ENTITY_TYPES, GIRCRedstoneMain.MODID);
 
     public static final RegistryObject<Block> RS_ACCEPTOR = internalRegisterBlock("acceptor",
-            () -> new BlockRedstoneAcceptor(BlockBehaviour.Properties.of() //TODO Material.METAL
+            () -> new BlockRedstoneAcceptor(BlockBehaviour.Properties.of() // TODO Material.METAL
                     .strength(1.5f, 6.0f).requiresCorrectToolForDrops()));
     public static final RegistryObject<Block> RS_EMITTER = internalRegisterBlock("emitter",
-            () -> new BlockRedstoneEmitter(BlockBehaviour.Properties.of() //TODO Material.METAL
+            () -> new BlockRedstoneEmitter(BlockBehaviour.Properties.of() // TODO Material.METAL
                     .strength(1.5f, 6.0f).requiresCorrectToolForDrops()));
     public static final RegistryObject<Block> RS_MULTI_EMITTER = internalRegisterBlock(
-            "multiemitter", () -> new BlockRedstoneMultiEmitter(BlockBehaviour.Properties
-                    .of().strength(1.5f, 6.0f).requiresCorrectToolForDrops())); //TODO Material.METAL
+            "multiemitter", () -> new BlockRedstoneMultiEmitter(BlockBehaviour.Properties.of()
+                    .strength(1.5f, 6.0f).requiresCorrectToolForDrops())); // TODO Material.METAL
 
     public static boolean acceptAcceptor(final Level level, final BlockPos pos) {
         return level.getBlockState(pos).getBlock() instanceof BlockRedstoneAcceptor;
     }
 
-    public static final RegistryObject<Item> RS_LINKER = ITEM_REGISTRY.register("linker",
-            () -> new Linkingtool(null, GIRCInit::acceptAcceptor));
+    public static final RegistryObject<Item> RS_LINKER =
+            ITEM_REGISTRY.register("linker", () -> new Linkingtool(null, GIRCInit::acceptAcceptor));
     public static final RegistryObject<Item> RS_MULTILINKER = ITEM_REGISTRY.register("multilinker",
             () -> new MultiLinkingTool(null, GIRCInit::acceptAcceptor));
     public static final RegistryObject<Item> REMOTE_ACTIVATOR = ITEM_REGISTRY.register("activator",
-            () -> new RemoteActivator());
+            () -> new RemoteActivator(null, GIRCInit::acceptAcceptor));
 
-    public static final RegistryObject<BlockEntityType<?>> EMITER_TILE = TILEENTITY_REGISTRY
-            .register("emitter", () -> BlockEntityType.Builder
+    public static final RegistryObject<BlockEntityType<?>> EMITER_TILE =
+            TILEENTITY_REGISTRY.register("emitter", () -> BlockEntityType.Builder
                     .of(TileRedstoneEmitter::new, RS_EMITTER.get()).build(null));
 
-    public static final RegistryObject<BlockEntityType<?>> MULTI_EMITER_TILE = TILEENTITY_REGISTRY
-            .register("multiemitter", () -> BlockEntityType.Builder
+    public static final RegistryObject<BlockEntityType<?>> MULTI_EMITER_TILE =
+            TILEENTITY_REGISTRY.register("multiemitter", () -> BlockEntityType.Builder
                     .of(TileRedstoneMultiEmitter::new, RS_MULTI_EMITTER.get()).build(null));
 
     private static final RegistryObject<Block> internalRegisterBlock(final String name,