+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.20.1 - 3.2.4]
+
+* feat: add item damage
+* fix: correct redstone update
+* fix: multilinking tooltag
+* ref: update linkable api
+* ref: update buildpipeline
+
 ## [1.20.1 - 3.2.2]
 
 * rem: debug output